Update VideoFrame2.py
**Second update
 CAN EXTRACT FRAMES AT 1 SECOND INTERVALS
Push to array and database pics to array for comparison
For every frame collected check against every pic in database and:
First greyscale both pics to compare
Resize so that they are the same dimensions
Use SSIM structural similarity testing from skiimages to find a comparison percentage
If percentage is below a tolerance, frame will be deleted from the EXPORT directory slow since it needs to run through every picture in the database for every frame. Could potentially just be my computer. Need to check